apim-traffic-tool/requirement.txt
certifi==2019.9.11
chardet==3.0.4
Faker==2.0.3
idna==2.8
numpy==1.17.3
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rstr==2.2.6
shyaml==0.6.1
six==1.12.0
text-unidecode==1.3
urllib3==1.25.6
